1) Add business-logic to the service layer; 2) Add Mapstruct dependency and mapper for a Place entity; 3) Fix some methods and interfaces and remove non usable methods

+++ src/main/java/com/example/geolocation_api/dao/dto/api/Location.java
@@ -13,10 +13,10 @@ import lombok.NoArgsConstructor;
 public class Location {
 
     @JsonProperty
-    private Float lat;
+    private Double lat;
 
     @JsonProperty
-    private Float lng;
+    private Double lng;
 
     @JsonProperty
     private String city;
